refactor: Extract UserImage from Whiteboard to own component
Also: Enable deleting Whiteboard images via Delete
