+++ src/components/Whiteboard/Whiteboard.js
@@ -1,7 +1,6 @@
 import BaseWindow from "../BaseWindow/BaseWindow";
 import React, { Component } from "react";
 import { Stage, Layer, Text, Image, Line, Transformer } from "react-konva";
-import useImage from "use-image"; // Konva-specific image hook
 
 // UI
 import ButtonGroup from "@mui/material/ButtonGroup";
@@ -14,6 +13,7 @@ import { useConfirm } from "material-ui-confirm";
 
 import ColorPalette from "./ColorPalette";
 import BrushPicker from "./BrushPicker";
+import UserImage from "./UserImage";
 
 // Icons
 import UndoIcon from "@mui/icons-material/Undo";
@@ -33,96 +33,6 @@ import CancelPresentationIcon from "@mui/icons-material/CancelPresentation";
 
 // CSS
 import "./Whiteboard.scss";
-
-function UserImage({
-  imageProps,
-  isSelected,
-  onSelect,
-  onChange,
-  isDraggable,
-  maxWidth,
-  maxHeight,
-}) {
-  const imageRef = React.useRef();
-  const imgTransformerRef = React.useRef();
-  const [image, status] = useImage(imageProps.url);
-
-  setImageSize();
-
-  React.useEffect(() => {
-    if (isSelected) {
-      // we need to attach transformer manually
-      imgTransformerRef.current.nodes([imageRef.current]);
-      imgTransformerRef.current.getLayer().batchDraw();
-    }
-  }, [isSelected]);
-
-  function setImageSize() {
-    if (status == "loaded") {
-      const ratio = image.width / image.height;
-      if (ratio > 1) {
-        image.height = maxHeight;
-        image.width = maxHeight * ratio;
-      } else {
-        image.width = maxWidth;
-        image.height = maxWidth / ratio;
-      }
-    }
-  }
-
-  return (
-    <React.Fragment>
-      <Image
-        image={image}
-        onClick={onSelect}
-        onTap={onSelect}
-        {...imageProps}
-        ref={imageRef}
-        draggable={isDraggable}
-        onDragEnd={(e) => {
-          onChange({
-            ...imageProps,
-            x: e.target.x(),
-            y: e.target.y(),
-          });
-        }}
-        onTransformEnd={(e) => {
-          // Transformer is changing scale of the node and not its width or height
-          // but in the store we have only width and height
-          // to match the data better we will reset scale on transform end
-          const node = imageRef.current;
-          const scaleX = node.scaleX();
-          const scaleY = node.scaleY();
-
-          // we will reset it back
-          node.scaleX(1);
-          node.scaleY(1);
-
-          onChange({
-            ...imageProps,
-            x: node.x(),
-            y: node.y(),
-            // set minimal value
-            width: Math.max(5, node.width() * scaleX),
-            height: Math.max(node.height() * scaleY),
-          });
-        }}
-      />
-      {isSelected && (
-        <Transformer
-          ref={imgTransformerRef}
-          boundBoxFunc={(oldBox, newBox) => {
-            // limit resize
-            if (newBox.width < 5 || newBox.height < 5) {
-              return oldBox;
-            }
-            return newBox;
-          }}
-        />
-      )}
-    </React.Fragment>
-  );
-}
 
 export default function Whiteboard({
   id,
@@ -265,7 +175,9 @@ export default function Whiteboard({
   };
 
   function handleStageWrapperKeyDown(event) {
-    if (event.keyCode === 8) {
+    // keyCode 8 = Backspace
+    // keyCode 46 = Delete
+    if (event.keyCode === 8 || event.keyCode === 46) {
       deleteSelectedImage();
     } else {
       return;
